pyproject.toml
[project]
name = "zalos"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.93.0",
    "python-pptx>=1.0.2",
    "requests>=2.32.4",
    "typing-extensions>=4.14.1",
]
